// src/lib/splitwise-api.d.ts
export interface paths {
// ...
    "/get_expense/{id}": {
        parameters: {
            query?: never;
            header?: never;
            path: {
                id: number;
            };
            cookie?: never;
        };
        /** Get expense information */
        get: {
            parameters: {
                query?: never;
                header?: never;
                path: {
                    id: number;
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description OK */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": {
                            expense?: components["schemas"]["expense"];
                        };
                    };
                };
                401: components["responses"]["unauthorized"];
                403: components["responses"]["forbidden"];
                404: components["responses"]["not_found"];
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/get_expenses": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** List the current user's expenses */
        get: {
            parameters: {
                query?: {
                    /** @description If provided, only expenses in that group will be returned, and `friend_id` will be ignored. */
                    group_id?: number;
                    /** @description ID of another user. If provided, only expenses between the current and provided user will be returned. */
                    friend_id?: number;
                    dated_after?: string;
                    dated_before?: string;
                    updated_after?: string;
                    updated_before?: string;
                    limit?: number;
                    offset?: number;
                };
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description OK */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": {
                            expenses?: components["schemas"]["expense"][];
                        };
                    };
                };
                401: components["responses"]["unauthorized"];
                403: components["responses"]["forbidden"];
                404: components["responses"]["not_found"];
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
// ...
        share: {
            user?: components["schemas"]["comment_user"];
            /** @example 491923 */
            user_id?: number;
            /** @example 8.99 */
            paid_share?: string;
            /** @example 4.5 */
            owed_share?: string;
            /** @example 4.49 */
            net_balance?: string;
        };
// ...
        expense: components["schemas"]["common"] & {
            /**
             * Format: int64
             * @example 51023
             */
            id?: number;
            /**
             * @description Null if the expense is not associated with a group.
             * @example 391
             */
            group_id?: number | null;
            /**
             * @description Null if the expense is not associated with a friendship.
             * @example 4818
             */
            friendship_id?: number | null;
            /** @example 491030 */
            expense_bundle_id?: number | null;
            /** @example Brunch */
            description?: string;
            /** @description Whether the expense recurs automatically */
            repeats?: boolean;
            /** @enum {string} */
            repeat_interval?: "never" | "weekly" | "fortnightly" | "monthly" | "yearly";
            /**
             * @description Whether a reminder will be sent to involved users in advance of the next occurrence of a recurring expense.
             *     Only applicable if the expense recurs.
             */
            email_reminder?: boolean;
            /**
             * @description Number of days in advance to remind involved users about the next occurrence of a new expense.
             *     Only applicable if the expense recurs.
             * @enum {integer|null}
             */
            email_reminder_in_advance?: null | -1 | 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 14;
            /** @description The date of the next occurrence of a recurring expense. Only applicable if the expense recurs. */
            next_repeat?: string | null;
            /** @description Also known as "notes." */
            details?: string | null;
            comments_count?: number;
            /** @description Whether this was a payment between users */
            payment?: boolean;
            /** @description If a payment was made via an integrated third party service, whether it was confirmed by that service. */
            transaction_confirmed?: boolean;
            /** @example 25.0 */
            cost?: string;
            /** @example USD */
            currency_code?: string;
            repayments?: {
                /**
                 * @description ID of the owing user
                 * @example 6788709
                 */
                from?: number;
                /**
                 * @description ID of the owed user
                 * @example 270896089
                 */
                to?: number;
                /** @example 25.0 */
                amount?: string;
            }[];
            /**
             * Format: date-time
             * @description The date and time the expense took place. May differ from `created_at`
             * @example 2012-05-02T13:00:00Z
             */
            date?: string;
            /**
             * Format: date-time
             * @description The date and time the expense was created on Splitwise
             * @example 2012-07-27T06:17:09Z
             */
            created_at?: string;
            created_by?: components["schemas"]["user"] & unknown;
            /**
             * Format: date-time
             * @description The last time the expense was updated.
             * @example 2012-12-23T05:47:02Z
             */
            updated_at?: string;
            updated_by?: components["schemas"]["user"] & unknown;
            /**
             * Format: date-time
             * @description If the expense was deleted, when it was deleted.
             * @example 2012-12-23T05:47:02Z
             */
            deleted_at?: string | null;
            deleted_by?: components["schemas"]["user"] & unknown;
            category?: {
                /** @example 5 */
                id?: number;
                /**
                 * @description Translated to the current user's locale
                 * @example Electricity
                 */
                name?: string;
            };
            receipt?: {
                /** @example https://splitwise.s3.amazonaws.com/uploads/expense/receipt/3678899/large_95f8ecd1-536b-44ce-ad9b-0a9498bb7cf0.png */
                large?: string | null;
                /** @example https://splitwise.s3.amazonaws.com/uploads/expense/receipt/3678899/95f8ecd1-536b-44ce-ad9b-0a9498bb7cf0.png */
                original?: string | null;
            };
            users?: components["schemas"]["share"][];
            comments?: components["schemas"]["comment"][];
        };
// ...
    };
// ...
}